Expose Result for Post-Processing
If a user needs to manipulate the files after upload, the resulting file paths should be exposed. These paths cannot be guessed from the request because the handler sanitizes the file name. This update also exposes a "complete" variable to easily determine if the file has been uploaded and combined.